fix: pylint errors
it seams that unused-arguments is not working as before. this commit
tries to use unused-arguments option in a more clearer way by following
the inherited method signature.

unused arguments which are in the signature of the parent method and are
mandatory there are not mentioned in the unused argument list of this
method by pylint. arguments which have a default value in the parent
method and are not used in the inherited method have to have the
unused-argument option to suppress the warning.